// frontend/src/app/model/item.model.ts
export class Item {
    id: number;
    name: string;
    code: string;
    partNumber: string;
    dbp:number;
    dbpCurrency: string;
    mrp:number;
    mrpCurrency: string;
    length: string;
    lengthUnit: string;
    attribute: string;
    unit: string;
    sku:number;
    creator: string;
    estimateShipTime: string;
    inStock:number;
    lang: string;
    longDesc: string;
    shippingCode: string;
    shortDesc: string;
    source: string;
    specialNotes: string;
    minOrderQty: number;
    maxOrderQty: number;
    largeImgAltText: string;
    largeImgName: string;
    largeImgUrl: string;
    smallImgAltText: string;
    smallImgLang: string;
    smallImgName: string;
    smallImgType: string;
    smallImgUrl: string;
    productId: number;
	particuler: string;
	varientcode: string;

    }
    